Redirect to main website
This repository represents an outdated naming and design for Mango Seed
Co. (previously named Mango Tree Co.). This is currently serving as the
endpoint for mangotreeco.com, but users should be redirected to the
actual website, mangoseedco.com.

To do this, an html file is added to handle redirecting users to the
correct endpoint. Further, the readme.md entrypoint is updated with a
permalink to the redirect html file instead of the old index.html file.